Discovery: Serve 'default' endpoint (#435)
This change adds a third endpoint called `default`. The discovery service will decide which privilege version of the topology file to serve, given the requester. This endpoint will be used by sciond. Thus, sciond on infra nodes must not be specifically configured to request the full topology.

Additionally, rename the endpoint `reduced` to `endhost` to make the intended use more explicit.

The changes to the open-source part of the code will be ported back in a follow-up PR.

fixes scionproto/scion#2441